+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 * Add a consumable resources concept to the Hosts Pool ([GH-205](https://github.com/ystia/yorc/issues/205))
 
+### ENHANCEMENTS
+
+* Yorc Tasks should have a error message when appropriate ([GH-613](https://github.com/ystia/yorc/issues/613))
+
 ### BUG FIXES
 
 * Yorc Bootstrap doesn't uninstall yorc binary ([GH-605](https://github.com/ystia/yorc/issues/605))